Fix relating configurable products to simple products, and fix "is_global = 1", "is_configurable = 1" attribute values displaying as empty on simple products.
Magento requires that configurable attributes are global, so they should be imported in the global scope. The exception is a few attributes like price that Magento has (erroneously?) set to "is_configurable = 1" even though they are not configurable attributes and they can be store-scoped. The previous logic was setting the scope of configurable attributes (except price) to be store-scoped. This is incorrect and causes the attribute to be imported in the store scope, which causes a failure to relate configurable products to simple products and causes the attribute value to display as empty on the simple product. The confusion may have been caused by the value of is_global, which is 1 for global scope, 0 for store scope, and 2 for website-scope. See Magmi_ProductImportEngine::getItemStoreIds.